+++ dlpxdbprofiler/__init__.py
@@ -8,4 +8,4 @@ Provides utilities to:
 for Oracle and MSSQL data sources.
 """
 
-__version__ = "1.1.0"
+__version__ = "1.2.0"